Pull out rec dropdown select and rec text inputs into their own files
